# contracts/agent.openapi.yaml
openapi: 3.1.0
info:
  title: Docklane Swarm Agent API
  version: 0.7.0
  description: Stable contract between the Docklane Control Plane and Go Agent.
servers:
  - url: https://127.0.0.1:9443
security:
  - mutualTLS: []

paths:
  /v1/health:
    get:
      operationId: health
      responses:
        '200':
          description: Agent is alive.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/cluster:
    get:
      operationId: inspectCluster
      responses:
        '200':
          description: Normalized Swarm cluster and node read model.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterResponse'

  /v1/nodes/{nodeId}:
    get:
      operationId: inspectNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
      responses:
        '200':
          description: Canonical node detail, non-terminal tasks, and affected services.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDetailResponse'

  /v1/nodes/{nodeId}/plan-labels:
    post:
      operationId: planNodeLabels
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeLabelPatchRequest'
      responses:
        '200':
          description: Deterministic full target label map and affected cluster services.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationPlan'
        '409':
          description: Expected node version is stale.

  /v1/nodes/{nodeId}/labels:
    post:
      operationId: updateNodeLabels
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeLabelsMutationRequest'
      responses:
        '200':
          description: Docker accepted the exact planned node label update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationResponse'
        '409':
          description: Node version, spec, or cluster service-set precondition failed.

  /v1/nodes/{nodeId}/plan-drain:
    post:
      operationId: planDrainNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodePlanRequest'
      responses:
        '200':
          description: Deterministic drain target and affected service IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationPlan'
        '409':
          description: Expected node version is stale.

  /v1/nodes/{nodeId}/drain:
    post:
      operationId: drainNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeMutationRequest'
      responses:
        '200':
          description: Docker accepted the exact planned node drain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationResponse'
        '409':
          description: Node version, spec, or affected task-set precondition failed.

  /v1/nodes/{nodeId}/plan-activate:
    post:
      operationId: planActivateNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodePlanRequest'
      responses:
        '200':
          description: Deterministic activate target.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationPlan'
        '409':
          description: Expected node version is stale.

  /v1/nodes/{nodeId}/activate:
    post:
      operationId: activateNode
      parameters:
        - $ref: '#/components/parameters/NodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeMutationRequest'
      responses:
        '200':
          description: Docker accepted the exact planned node activation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeMutationResponse'
        '409':
          description: Node version or spec precondition failed.

  /v1/services:
    get:
      operationId: listServices
      responses:
        '200':
          description: Normalized Swarm services.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceSummary'

  /v1/services/{serviceId}:
    get:
      operationId: inspectService
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Normalized service and tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'

  /v1/services/{serviceId}/tasks:
    get:
      operationId: listServiceTasks
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      responses:
        '200':
          description: Normalized service tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskSummary'

  /v1/services/{serviceId}/capacity-check:
    post:
      operationId: checkServiceCapacity
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapacityCheckRequest'
      responses:
        '200':
          description: Conservative Swarm capacity estimate for the requested target replicas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapacityCheckResponse'
        '409':
          description: Expected service version is stale.

  /v1/services/{serviceId}/plan-scale:
    post:
      operationId: planScaleService
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScalePlanRequest'
      responses:
        '200':
          description: Canonical service identity and deterministic target spec.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMutationPlan'
        '409':
          description: Expected service version is stale.

  /v1/services/{serviceId}/plan-restart:
    post:
      operationId: planRestartService
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestartPlanRequest'
      responses:
        '200':
          description: Canonical service identity and deterministic target spec.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMutationPlan'
        '409':
          description: Expected service version is stale.

  /v1/services/{serviceId}/scale:
    post:
      operationId: scaleService
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScaleMutationRequest'
      responses:
        '200':
          description: Docker accepted the exact planned service update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMutationResponse'
        '400':
          description: Invalid service mode or request.
        '409':
          description: Version or spec precondition failed.

  /v1/services/{serviceId}/restart:
    post:
      operationId: restartService
      parameters:
        - $ref: '#/components/parameters/ServiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestartMutationRequest'
      responses:
        '200':
          description: Docker accepted the exact planned forced update.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceMutationResponse'
        '409':
          description: Version or spec precondition failed.

components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS

  parameters:
    NodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: string
        minLength: 1

    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1

  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, component]
      properties:
        status:
          type: string
          const: ok
        component:
          type: string
          const: docklane-agent

    ManagerQuorum:
      type: object
      additionalProperties: false
      required: [total, reachable, required, available, leaderCount]
      properties:
        total: { type: integer, minimum: 0 }
        reachable: { type: integer, minimum: 0 }
        required: { type: integer, minimum: 0 }
        available: { type: boolean }
        leaderCount: { type: integer, minimum: 0 }

    NodeSummary:
      type: object
      additionalProperties: false
      required:
        [id, version, specHash, hostname, address, role, availability, state, manager, leader, nanoCpus, memoryBytes]
      properties:
        id: { type: string }
        version: { type: integer, minimum: 0 }
        specHash: { type: string, minLength: 1 }
        hostname: { type: string }
        address: { type: string }
        role: { type: string }
        availability: { type: string }
        state: { type: string }
        message: { type: string }
        manager: { type: boolean }
        leader: { type: boolean }
        reachability: { type: string }
        engineVersion: { type: string }
        nanoCpus: { type: integer }
        memoryBytes: { type: integer }
        labels:
          type: object
          additionalProperties:
            type: string

    ClusterSummary:
      type: object
      additionalProperties: false
      required: [id, dockerVersion, apiVersion, createdAt, updatedAt, managers]
      properties:
        id: { type: string }
        dockerVersion: { type: string }
        apiVersion: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        managers:
          $ref: '#/components/schemas/ManagerQuorum'

    ClusterResponse:
      type: object
      additionalProperties: false
      required: [cluster, nodes]
      properties:
        cluster:
          $ref: '#/components/schemas/ClusterSummary'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeSummary'

    NodeDetailResponse:
      type: object
      additionalProperties: false
      required: [node, tasks, serviceIds]
      properties:
        node:
          $ref: '#/components/schemas/NodeSummary'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummary'
        serviceIds:
          type: array
          items: { type: string }

    NodeLabelPatchRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, set, remove]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        set:
          type: object
          additionalProperties:
            type: string
        remove:
          type: array
          items: { type: string }

    NodeLabelsMutationRequest:
      type: object
      additionalProperties: false
      required:
        [expectedVersion, expectedSpecHash, targetSpecHash, expectedServiceIds, targetLabels]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        expectedSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }
        expectedServiceIds:
          type: array
          items: { type: string }
        targetLabels:
          type: object
          additionalProperties:
            type: string

    NodePlanRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 0 }

    NodeMutationRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, expectedSpecHash, targetSpecHash]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        expectedSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }
        expectedServiceIds:
          type: array
          items: { type: string }

    NodeMutationPlan:
      type: object
      additionalProperties: false
      required:
        [nodeId, version, beforeSpecHash, targetSpecHash, targetAvailability, affectedServiceIds]
      properties:
        nodeId: { type: string }
        version: { type: integer, minimum: 0 }
        beforeSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }
        targetAvailability:
          type: string
          enum: [drain, active, pause]
        affectedServiceIds:
          type: array
          items: { type: string }
        targetLabels:
          type: object
          additionalProperties:
            type: string

    NodeMutationResponse:
      type: object
      additionalProperties: false
      required: [nodeId, version, targetSpecHash, targetAvailability]
      properties:
        nodeId: { type: string }
        version: { type: integer, minimum: 0 }
        targetSpecHash: { type: string, minLength: 1 }
        targetAvailability:
          type: string
          enum: [drain, active]

    ServiceSummary:
      type: object
      additionalProperties: false
      required:
        [id, name, version, specHash, forceUpdate, mode, desiredReplicas, runningReplicas, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        version: { type: integer, minimum: 0 }
        specHash: { type: string, minLength: 1 }
        forceUpdate: { type: integer, minimum: 0 }
        image: { type: string }
        mode: { type: string }
        desiredReplicas: { type: integer, minimum: 0 }
        runningReplicas: { type: integer, minimum: 0 }
        updateState: { type: string }
        updateMessage: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    TaskSummary:
      type: object
      additionalProperties: false
      required:
        [id, serviceId, slot, desiredState, state, forceUpdate, timestamp]
      properties:
        id: { type: string }
        serviceId: { type: string }
        slot: { type: integer }
        nodeId: { type: string }
        desiredState: { type: string }
        state: { type: string }
        forceUpdate: { type: integer, minimum: 0 }
        message: { type: string }
        error: { type: string }
        containerId: { type: string }
        image: { type: string }
        timestamp: { type: string, format: date-time }

    ServiceDetailResponse:
      type: object
      additionalProperties: false
      required: [service, tasks]
      properties:
        service:
          $ref: '#/components/schemas/ServiceSummary'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskSummary'

    CapacityCheckRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, targetReplicas, includeUpdateOverlap]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        targetReplicas: { type: integer, minimum: 0 }
        includeUpdateOverlap: { type: boolean }

    CapacityResource:
      type: object
      additionalProperties: false
      required: [nanoCpus, memoryBytes]
      properties:
        nanoCpus: { type: integer, minimum: 0 }
        memoryBytes: { type: integer, minimum: 0 }

    CapacityNode:
      type: object
      additionalProperties: false
      required:
        [nodeId, hostname, availableNanoCpus, availableMemoryBytes, existingServiceTasks, maxAdditionalReplicas]
      properties:
        nodeId: { type: string }
        hostname: { type: string }
        availableNanoCpus: { type: integer, minimum: 0 }
        availableMemoryBytes: { type: integer, minimum: 0 }
        existingServiceTasks: { type: integer, minimum: 0 }
        maxAdditionalReplicas: { type: integer, minimum: 0 }

    CapacityCheckResponse:
      type: object
      additionalProperties: false
      required:
        [serviceId, status, currentReplicas, targetReplicas, updateOverlapReplicas, requiredAdditionalReplicas, schedulableAdditionalReplicas, unplacedReplicas, eligibleNodeCount, reservation, reasons, unsupportedConstraints, nodes]
      properties:
        serviceId: { type: string }
        status:
          type: string
          enum: [SUFFICIENT, INSUFFICIENT, UNKNOWN]
        currentReplicas: { type: integer, minimum: 0 }
        targetReplicas: { type: integer, minimum: 0 }
        updateOverlapReplicas: { type: integer, minimum: 0 }
        requiredAdditionalReplicas: { type: integer, minimum: 0 }
        schedulableAdditionalReplicas: { type: integer, minimum: 0 }
        unplacedReplicas: { type: integer, minimum: 0 }
        eligibleNodeCount: { type: integer, minimum: 0 }
        reservation:
          $ref: '#/components/schemas/CapacityResource'
        reasons:
          type: array
          items: { type: string }
        unsupportedConstraints:
          type: array
          items: { type: string }
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/CapacityNode'

    ScalePlanRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, replicas]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        replicas: { type: integer, minimum: 0 }

    RestartPlanRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, minimum: 0 }

    ScaleMutationRequest:
      type: object
      additionalProperties: false
      required:
        [expectedVersion, expectedSpecHash, targetSpecHash, replicas]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        expectedSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }
        replicas: { type: integer, minimum: 0 }

    RestartMutationRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, expectedSpecHash, targetSpecHash]
      properties:
        expectedVersion: { type: integer, minimum: 0 }
        expectedSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }

    ServiceMutationPlan:
      type: object
      additionalProperties: false
      required:
        [serviceId, version, beforeSpecHash, targetSpecHash, targetForceUpdate]
      properties:
        serviceId: { type: string }
        version: { type: integer, minimum: 0 }
        beforeSpecHash: { type: string, minLength: 1 }
        targetSpecHash: { type: string, minLength: 1 }
        targetForceUpdate: { type: integer, minimum: 0 }
        targetReplicas: { type: integer, minimum: 0 }

    ServiceMutationResponse:
      type: object
      additionalProperties: false
      required:
        [serviceId, version, targetSpecHash, targetForceUpdate]
      properties:
        serviceId: { type: string }
        version: { type: integer, minimum: 0 }
        targetSpecHash: { type: string, minLength: 1 }
        targetForceUpdate: { type: integer, minimum: 0 }
        warnings:
          type: array
          items: { type: string }